+++ lib/capybara/webkit/browser.rb
@@ -39,12 +39,8 @@ module Capybara::Webkit
     end
 
     def console_messages
-      command("ConsoleMessages").split("\n").map do |messages|
-        parts = messages.split("|", 3)
-        message = parts.pop.gsub("\\n", "\n")
-        { :source => parts.first, :message => message }.tap do |message|
-          message[:line_number] = Integer(parts[1]) if parts[1]
-        end
+      JSON.parse(command("ConsoleMessages")).map do |message|
+        message.inject({}) { |m,(k,v)| m.merge(k.to_sym => v) }
       end
     end
 
@@ -55,15 +51,15 @@ module Capybara::Webkit
     end
 
     def alert_messages
-      command("JavascriptAlertMessages").split("\n")
+      JSON.parse(command("JavascriptAlertMessages"))
     end
 
     def confirm_messages
-      command("JavascriptConfirmMessages").split("\n")
+      JSON.parse(command("JavascriptConfirmMessages"))
     end
 
     def prompt_messages
-      command("JavascriptPromptMessages").split("\n")
+      JSON.parse(command("JavascriptPromptMessages"))
     end
 
     def response_headers
@@ -204,7 +200,7 @@ module Capybara::Webkit
       elsif result != 'ok'
         case response = read_response
         when "timeout"
-          raise Capybara::TimeoutError, "Request timed out after #{timeout_seconds}"
+          raise Timeout::Error, "Request timed out after #{timeout_seconds}"
         else
           raise InvalidResponseError, response
         end