+++ spec/spec_helper.rb
@@ -33,7 +33,7 @@ end
 
 RSpec.configure do |c|
   c.filter_run_excluding :skip_on_windows => !(RbConfig::CONFIG['host_os'] =~ /mingw32/).nil?
-  c.before { Capybara.app_host = nil }
+  Capybara::SpecHelper.configure(c)
 end
 
 def with_env_vars(vars)