+++ src/Command.h
@@ -3,6 +3,7 @@
 
 #include <QObject>
 #include "Response.h"
+#include <QString>
 
 class Command : public QObject {
   Q_OBJECT
@@ -11,6 +12,11 @@ class Command : public QObject {
     Command(QObject *parent = 0);
     virtual void start() = 0;
     virtual QString toString() const;
+
+  protected:
+    void emitFinished(bool success);
+    void emitFinished(bool success, QString message);
+    void emitFinished(bool success, QByteArray message);
 
   signals:
     void finished(Response *response);