+++ src/ConsoleMessages.cpp
@@ -1,11 +1,14 @@
 #include "ConsoleMessages.h"
 #include "WebPage.h"
 #include "WebPageManager.h"
+#include "JsonSerializer.h"
 
 ConsoleMessages::ConsoleMessages(WebPageManager *manager, QStringList &arguments, QObject *parent) : SocketCommand(manager, arguments, parent) {
 }
 
 void ConsoleMessages::start() {
-  emit finished(new Response(true, page()->consoleMessages()));
+  JsonSerializer serializer;
+  QByteArray json = serializer.serialize(page()->consoleMessages());
+  emitFinished(true, json);
 }