+++ src/Evaluate.h
@@ -8,13 +8,5 @@ class Evaluate : public SocketCommand {
   public:
     Evaluate(WebPageManager *, QStringList &arguments, QObject *parent = 0);
     virtual void start();
-
-  private:
-    void addVariant(QVariant &object);
-    void addString(QString &string);
-    void addArray(QVariantList &list);
-    void addMap(QVariantMap &map);
-
-    QString m_buffer;
 };