+++ src/FrameFocus.cpp
@@ -51,7 +51,7 @@ void FrameFocus::focusId(QString name) {
 
 void FrameFocus::focusParent() {
   if (page()->currentFrame()->parentFrame() == 0) {
-    emit finished(new Response(false, QString("Already at parent frame.")));
+    emitFinished(false, QString("Already at parent frame."));
   } else {
     page()->currentFrame()->parentFrame()->setFocus();
     success();
@@ -59,9 +59,9 @@ void FrameFocus::focusParent() {
 }
 
 void FrameFocus::frameNotFound() {
-  emit finished(new Response(false, QString("Unable to locate frame. ")));
+  emitFinished(false, QString("Unable to locate frame. "));
 }
 
 void FrameFocus::success() {
-  emit finished(new Response(true));
+  emitFinished(true);
 }