+++ src/JavascriptAlertMessages.cpp
@@ -1,10 +1,13 @@
 #include "JavascriptAlertMessages.h"
 #include "WebPage.h"
 #include "WebPageManager.h"
+#include "JsonSerializer.h"
 
 JavascriptAlertMessages::JavascriptAlertMessages(WebPageManager *manager, QStringList &arguments, QObject *parent) : SocketCommand(manager, arguments, parent) {}
 
 void JavascriptAlertMessages::start()
 {
-  emit finished(new Response(true, page()->alertMessages()));
+  JsonSerializer serializer;
+  QByteArray json = serializer.serialize(page()->alertMessages());
+  emitFinished(true, json);
 }