+++ src/SetUrlBlacklist.cpp
@@ -10,6 +10,6 @@ SetUrlBlacklist::SetUrlBlacklist(WebPageManager *manager, QStringList &arguments
 void SetUrlBlacklist::start() {
   NetworkAccessManager* networkAccessManager = page()->networkAccessManager();
   networkAccessManager->setUrlBlacklist(arguments());
-  emit finished(new Response(true));
+  emitFinished(true);
 }